requirements.txt
pytest
blinker
detach
pyserial
sphinx-me
